# documentation/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Intra API Documentation",
    "description": "Welcome to the API Documentation for the Intra application. This API provides a set of endpoints to interact with various features of the application. Below are the different categories and their corresponding endpoints:\n"
  },
  "servers": [
    {
      "url": "https://le-lien-de-lapi"
    }
  ],
...
  "paths": {
...
    "/UNITS/{ID}": {
      "get": {
        "summary": "Get unit details",
        "description": "Retrieves the details of a specific unit.",
        "tags": [
          "Units"
        ],
        "parameters": [
          {
            "name": "ID",
            "in": "path",
            "description": "The ID of the unit.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "APIKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "API-Key",
        "description": "Provide your API key as a header parameter. You can obtain the API key by authenticating with the application.\n"
      }
    }
  },
  "security": [
    {
      "APIKeyAuth": []
    }
  ]
}